Refactor: Course Card 리팩토링 #80
- useCountDown 분리
- MeatBall 분리
- 기타 함수 분리
